+++ modules/ModuleNewsCategories.php
@@ -112,7 +112,7 @@ class ModuleNewsCategories extends \ModuleNews
 		while ($objCategories->next())
 		{
 			$arrCategories[$objCategories->id] = $objCategories->row();
-			$arrCategories[$objCategories->id]['class'] = trim(((++$count == 1) ? ' first' : '') . (($count == $total) ? ' last' : '') . ((($count % 2) == 0) ? ' odd' : ' even'));
+			$arrCategories[$objCategories->id]['class'] = 'news_category_' . $objCategories->id . ((++$count == 1) ? ' first' : '') . (($count == $total) ? ' last' : '') . ((($count % 2) == 0) ? ' odd' : ' even');
 			$arrCategories[$objCategories->id]['title'] = specialchars($objCategories->title);
 			$arrCategories[$objCategories->id]['href'] = ampersand(sprintf($strUrl, ($GLOBALS['TL_CONFIG']['disableAlias'] ? $objCategories->id : $objCategories->alias)));
 			$arrCategories[$objCategories->id]['link'] = $objCategories->title;